proper rule inclusion
